// src/pages/edgeCloudMessage/ruleRoute/data.d.ts
export type listType = {
  name: string;
  uid: string;
  creationTimestamp: string
};
export type formType = {
  name: string;
  source: string;
  sourceResource: string;
  target: string;
  targetResource: string;
  description: string
}
export type DeptType = {
  deptId: number;
  parentId: number;
  ancestors: string;
  deptName: string;
  orderNum: number;
  leader: string;
  phone: string;
  email: string;
  status: string;
  delFlag: string;
  createBy: string;
  createTime: Date;
  updateBy: string;
  updateTime: Date;
};

export type DeptListPagination = {
  total: number;
  pageSize: number;
  current: number;
};

export type DeptListData = {
  list: DeptType[];
  pagination: Partial<DeptListPagination>;
};

export type DeptListParams = {
  deptId?: string;
  parentId?: string;
  ancestors?: string;
  deptName?: string;
  orderNum?: string;
  leader?: string;
  phone?: string;
  email?: string;
  status?: string;
  delFlag?: string;
  createBy?: string;
  createTime?: string;
  updateBy?: string;
  updateTime?: string;
  pageSize?: string;
  currentPage?: string;
  filter?: string;
  sorter?: string;
};